Improved code clarity and deleted unnecessary parts
